Added warning and pause/win/lose menu dialogue
Warning dialogue is high priority dialogue that is played immediately.
It is useful to notify the user if towers have been taken or if the
king's health is getting low.

Realtime dialogue can be played when the game is paused or when
displaying the win and lose screen.

Random chatter has not yet been implemented.